seva-launcher: Break out docker-compose components
